KTX2Loader: Prefer traditional for loop
This PR addresses a syntax issue that was causing problems with Babel's conversion of the original for...of loop. To ensure compatibility and avoid issues during transpilation, the loop has been refactored to use a traditional for loop with an index variable.

This change ensures broader compatibility and smoother integration with Babel if someone wants to embed the KTX2Loader source.